NewUpdation

+++ src/resolvers/Mutations.js
@@ -1,6 +1,6 @@
 import { generateOtp } from "../util/otp.js";
 import { verifyOTP } from "../util/otp.js";
-import pkg from 'mongodb';
+import pkg from "mongodb";
 import ObjectID from "mongodb";
 const { Long } = pkg;
 import password_1 from "@accounts/password";
@@ -38,11 +38,11 @@ export default {
     const { users } = collections;
     const email = args.email;
     const phone = args.phone;
-    console.log(args)
+    console.log(args);
 
     let userExist = await users.findOne({ phone: phone });
     if (!userExist) {
-      userExist = await users.findOne({ 'emails.0.address': email });
+      userExist = await users.findOne({ "emails.0.address": email });
     }
 
     console.log("userExist");
@@ -58,14 +58,16 @@ export default {
     return verifyOTP(args.phone, args.otp, context);
   },
   resetPassword: async (_, { token, newPassword }, { injector, infos }) => {
-    return injector.get(password_1.AccountsPassword).resetPassword(token, newPassword, infos);
+    return injector
+      .get(password_1.AccountsPassword)
+      .resetPassword(token, newPassword, infos);
   },
 
   sendResetPasswordEmail: async (_, { email }, { injector }, ctx) => {
-    console.log(ctx)
+    console.log(ctx);
     const { backgroundJobs, collections } = ctx;
-    console.log(collections)
-    console.log(backgroundJobs)
+    console.log(collections);
+    console.log(backgroundJobs);
     const accountsServer = injector.get(server_1.AccountsServer);
     const accountsPassword = injector.get(password_1.AccountsPassword);
     // const randomNumber = Math.floor(Math.random() * 1000000);
@@ -79,13 +81,14 @@ export default {
       // await accountsPassword.sendResetPasswordEmail(user.id, otp);
       // await accountsPassword.sendResetPasswordEmail(user, otp);
       await accountsPassword.sendResetPasswordEmail(email);
-    }
-    catch (error) {
+    } catch (error) {
       // If ambiguousErrorMessages is true,
       // to prevent user enumeration we fail silently in case there is no user attached to this email
-      if (accountsServer.options.ambiguousErrorMessages &&
+      if (
+        accountsServer.options.ambiguousErrorMessages &&
         error instanceof server_1.AccountsJsError &&
-        error.code === password_1.SendResetPasswordEmailErrors.UserNotFound) {
+        error.code === password_1.SendResetPasswordEmailErrors.UserNotFound
+      ) {
         return null;
       }
       throw error;
@@ -130,29 +133,31 @@ export default {
       const createdUser = await accountsServer.findUserById(userId);
       // If we are here - user must be created successfully
       // Explicitly saying this to Typescript compiler
-      const loginResult = await accountsServer.loginWithUser(createdUser, infos);
+      const loginResult = await accountsServer.loginWithUser(
+        createdUser,
+        infos
+      );
       await generateOtp(user.phone);
       return {
         userId,
         // loginResult,
       };
-    }
-    else {
+    } else {
       if (!ctx.authToken) {
-        throw new Error("Unauthorized")
+        throw new Error("Unauthorized");
       }
       if (ctx.user === undefined || ctx.user === null) {
-        throw new Error("Unauthorized")
+        throw new Error("Unauthorized");
       }
-      console.log(ctx.user.UserRole)
-      console.log(user)
-      console.log(user.UserRole)
+      console.log(ctx.user.UserRole);
+      console.log(user);
+      console.log(user.UserRole);
 
       // Check Permission for create user
-      console.log(ctx.user.UserRole)
-      console.log(user.UserRole)
-      const UserPermission = canCreateUser(ctx.user.UserRole, user.UserRole)
-      console.log(UserPermission)
+      console.log(ctx.user.UserRole);
+      console.log(user.UserRole);
+      const UserPermission = canCreateUser(ctx.user.UserRole, user.UserRole);
+      console.log(UserPermission);
       if (UserPermission) {
         // Allow user creation
         try {
@@ -172,19 +177,55 @@ export default {
         }
         if (!accountsServer.options.enableAutologin) {
           return {
-            userId: accountsServer.options.ambiguousErrorMessages ? null : userId,
+            userId: accountsServer.options.ambiguousErrorMessages
+              ? null
+              : userId,
           };
         }
-        // if (userId) {
-        //         const accountAdded = await Accounts.insertOne({ _id: userId, firstName: user.firstName, lastName: user.lastName, name: user.firstName + " " + user.lastName, phone: user.phone })
-
-        // }
+        if (userId) {
+          const account = {
+            "_id": userId,
+            "acceptsMarketing": false,
+            "emails": [
+              {
+                "address": user.email,
+                "verified": false,
+                "provides": "default"
+              }
+            ],
+            "groups": [],
+            "name": null,
+            "profile": {
+              firstName: user.firstName,
+              lastName: user.lastName,
+              dob: user.dob,
+              phone: user.phone,
+            },
+            "shopId": null,
+            "state": "new",
+            "userId": userId,
+            "UserRole": user.UserRole,
+          }
+          // const accountAdded = await Accounts.insertOne({
+          //   _id: userId,
+          //   firstName: user.firstName,
+          //   lastName: user.lastName,
+          //   name: user.firstName + " " + user.lastName,
+          //   phone: user.phone,
+          //   UserRole: user.UserRole
+          // });
+          const accountAdded = await Accounts.insertOne(account);
+          console.log(accountAdded)
+        }
         // When initializing AccountsServer we check that enableAutologin and ambiguousErrorMessages options
         // are not enabled at the same time
         const createdUser = await accountsServer.findUserById(userId);
         // If we are here - user must be created successfully
         // Explicitly saying this to Typescript compiler
-        const loginResult = await accountsServer.loginWithUser(createdUser, infos);
+        const loginResult = await accountsServer.loginWithUser(
+          createdUser,
+          infos
+        );
         await generateOtp(user.phone);
         return {
           userId,
@@ -192,18 +233,22 @@ export default {
         };
       } else {
         // Deny user creation
-        throw new Error("Unauthorized")
+        throw new Error("Unauthorized");
       }
-
-
     }
   },
-  changePassword: async (_, { oldPassword, newPassword }, { user, injector }) => {
+  changePassword: async (
+    _,
+    { oldPassword, newPassword },
+    { user, injector }
+  ) => {
     if (!(user && user.id)) {
-      throw new Error('Unauthorized');
+      throw new Error("Unauthorized");
     }
     const userId = user.id;
-    await injector.get(password_1.AccountsPassword).changePassword(userId, oldPassword, newPassword);
+    await injector
+      .get(password_1.AccountsPassword)
+      .changePassword(userId, oldPassword, newPassword);
     return null;
   },
   async createUserWithOtp(_, { user }, ctx) {
@@ -216,7 +261,6 @@ export default {
 
     try {
       userId = await accountsPassword.createUser(user);
-
     } catch (error) {
       // If ambiguousErrorMessages is true we obfuscate the email or username already exist error
       // to prevent user enumeration during user creation
@@ -236,35 +280,35 @@ export default {
       };
     }
 
-
-    const adminCount = await Accounts.findOne({ "adminUIShopIds.0": { $ne: null } });
+    const adminCount = await Accounts.findOne({
+      "adminUIShopIds.0": { $ne: null },
+    });
     console.log("adminCount", adminCount);
     if (userId && adminCount?._id) {
-      console.log("user", user)
+      console.log("user", user);
       const account = {
-        "_id": userId,
-        "acceptsMarketing": false,
-        "emails": [
+        _id: userId,
+        acceptsMarketing: false,
+        emails: [
           {
-            "address": user.email,
-            "verified": false,
-            "provides": "default"
-          }
+            address: user.email,
+            verified: false,
+            provides: "default",
+          },
         ],
-        "groups": [],
-        "name": null,
-        "profile": {
+        groups: [],
+        name: null,
+        profile: {
           firstName: user.firstName,
           lastName: user.lastName,
           dob: user.dob,
           phone: user.phone,
         },
-        "shopId": null,
-        "state": "new",
-        "userId": userId
-      }
+        shopId: null,
+        state: "new",
+        userId: userId,
+      };
       const accountAdded = await Accounts.insertOne(account);
-
     }
     // When initializing AccountsServer we check that enableAutologin and ambiguousErrorMessages options
     // are not enabled at the same time
@@ -282,7 +326,7 @@ export default {
     const { serviceName, params } = args;
     const { injector, infos, collections } = ctx;
     const { users } = collections;
-    console.log("authenticate")
+    console.log("authenticate");
     const authenticated = await injector
       .get(server_1.AccountsServer)
       .loginWithService(serviceName, params, infos);
@@ -312,7 +356,7 @@ export default {
       const authenticated = await injector
         .get(server_1.AccountsServer)
         .loginWithService(serviceName, params, infos);
-      console.log("authenticated", authenticated)
+      console.log("authenticated", authenticated);
       return authenticated;
     }
   },
@@ -324,14 +368,14 @@ export default {
       throw new Error("Unauthorized access. Please login first");
     }
     if (!context.authToken) {
-      throw new Error('UnAuthorization!');
+      throw new Error("Unauthorized");
     }
+
     const { users, Accounts } = context.collections;
     const { userId } = args;
 
-
     const query = { _id: ObjectID(userId) };
-    console.log('query:', query);
+    console.log("query:", query);
 
     // const usersList = await users.find().toArray();
     // usersList.forEach(user => {
@@ -339,29 +383,38 @@ export default {
     //   return user.phone
     // });
 
-    const userResponse = await users.findOne({ _id: args.userId });
     const userAccountResponse = await Accounts.findOne({ _id: args.userId });
-    console.log("User Response : ", userResponse)
+    if (userAccountResponse && userAccountResponse.adminUIShopIds && userAccountResponse.adminUIShopIds.length > 0) {
+      throw new Error("Cannot delete Super Admin");
+    }
+    const userResponse = await users.findOne({ _id: args.userId });
+    console.log("User Response : ", userResponse);
     if (!userResponse) {
       throw new Error(`User not found`);
     }
-    const UserPermission = await canCreateUser(context.user.UserRole, userResponse.UserRole)
-    console.log(UserPermission)
+    const UserPermission = await canCreateUser(
+      context.user.UserRole,
+      userResponse.UserRole
+    );
+    console.log(UserPermission);
     if (UserPermission) {
       const deleteResult = await users.deleteOne({ _id: args.userId });
-      console.log(deleteResult)
-      const deleteAccountResult = await Accounts.deleteOne({ _id: args.userId });
-      if (deleteResult.deletedCount > 0 || deleteAccountResult.deletedCount > 0) {
+      console.log(deleteResult);
+      const deleteAccountResult = await Accounts.deleteOne({
+        _id: args.userId,
+      });
+      if (
+        deleteResult.deletedCount > 0 ||
+        deleteAccountResult.deletedCount > 0
+      ) {
         return true; // User deleted successfully, return true
+      } else {
+        return false;
       }
-      else {
-        return false
-      }
-    }
-    else {
+    } else {
       // Deny user creation
-      throw new Error("Unauthorized")
+      throw new Error("Unauthorized");
     }
     return UserPermission;
-  }
+  },
 };
